+++ src/scip/cons_expr_pow.c
@@ -429,7 +429,7 @@ SCIP_DECL_CONSEXPR_EXPRHASH(hashPow)
    assert(expr2key != NULL);
    assert(hashkey != NULL);
 
-   *hashkey = POW_HASHKEY;
+   *hashkey = (unsigned int)POW_HASHKEY;
 
    assert(SCIPhashmapExists(expr2key, (void*)SCIPgetConsExprExprChildren(expr)[0]));
    childhash = (unsigned int)(size_t)SCIPhashmapGetImage(expr2key, SCIPgetConsExprExprChildren(expr)[0]);
